Move profile window easter egg to the render stats window.
